+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
     "antd": "^4.19.5",
+    "bootstrap-material-design": "^4.1.3",
     "mdbreact": "^5.2.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
