+++ src/App.js
@@ -3,12 +3,12 @@ import { Routes, Route } from "react-router-dom";
 import Login from "./pages/auth/Login";
 import Registration from "./pages/auth/Register";
 import Home from "./pages/Home";
-import Header from "./components/nav/Header";
+import AppHeader from "./components/nav/AppHeader";
 
 function App() {
   return (
     <>
-      <Header />
+      <AppHeader />
       <Routes>
         <Route index element={<Home />} />
         <Route path='/login' element={<Login />} />
