+++ src/pages/Home.js
@@ -1,7 +1,11 @@
 import React from "react";
 
 const Home = () => {
-  return <div>Home</div>;
+  return (
+    <>
+      <div className='float-end'>AKila</div>
+    </>
+  );
 };
 
 export default Home;
